Move properties to TS
